# ibm/openapi.yaml
openapi: 3.0.3
info:
  title: WaveGuard Orchestrate Tools
  version: 0.1.0
  description: >
    Read-only WaveGuard operations exposed as IBM watsonx Orchestrate tools.
    The deterministic backend owns severity. These tools report it and ask
    Granite to explain it; none of them change severity or trigger hardware.
servers:
  # watsonx Orchestrate must reach the backend over HTTPS. During the demo,
  # expose the local API with a tunnel (for example `ngrok http 8000`) and
  # replace this URL before importing the file.
  - url: https://REPLACE-WITH-TUNNEL.example.com
paths:
...
  /api/events/{event_id}/explain:
    post:
      operationId: explainEvent
      summary: Explain a stored event with IBM Granite
      description: >
        Sends the stored event (never raw sensor text) to IBM Granite and returns
        an operator summary, evidence, recommended actions, and a public-warning
        draft. Severity in the response is copied from the detector.
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
          example: evt-1a2b3c4d5e
      responses:
        "200":
          description: Granite explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplanationResult"
        "404":
          description: Event not found
        "502":
          description: IBM watsonx.ai rejected or failed the request
        "503":
          description: Granite credentials are not configured
components:
  schemas:
    Severity:
      type: string
      enum: [safe, watch, warning]
...
    ExplanationResult:
      type: object
      properties:
        event_id:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        classification:
          type: string
        source:
          type: string
          enum: [granite, prerecorded]
        model_id:
          type: string
          nullable: true
        generated_at:
          type: string
          format: date-time
        explanation:
          type: object
          properties:
            summary:
              type: string
            evidence:
              type: array
              items:
                type: string
            recommended_actions:
              type: array
              items:
                type: string
            public_warning:
              type: string